requirements.txt
Django==4.2.7
channels==4.0.0
Pillow==11.0.0
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
